backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.2.1
python-multipart==0.0.9
google-generativeai==0.8.3

